Moved node localStorage folder to user directory
