+++ src/node/util/util.ts
@@ -1,0 +1,28 @@
+/**
+ * Utilities useful for node processes
+ */
+import { URL } from 'url';
+import path from 'path';
+
+export function resolveHtmlPath(htmlFileName: string) {
+  if (process.env.NODE_ENV === 'development') {
+    const port = process.env.PORT || 1212;
+    const url = new URL(`http://localhost:${port}`);
+    url.pathname = htmlFileName;
+    return url.href;
+  }
+  return `file://${path.resolve(__dirname, '../renderer/', htmlFileName)}`;
+}
+
+/**
+ * Gets the platform-specific user folder for this application
+ * Thanks to Luke at https://stackoverflow.com/a/26227660
+ */
+export function getUserDir(): string {
+  return `${
+    process.env.APPDATA ||
+    (process.platform === 'darwin'
+      ? `${process.env.HOME}/Library/Preferences`
+      : `${process.env.HOME}/.local/share`)
+  }/paranext-core`;
+}
